desafio_com_relacionamento/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
distlib==0.3.6
Django==4.1.7
djangorestframework==3.14.0
filelock==3.10.7
pbr==5.11.1
platformdirs==3.2.0
psycopg2==2.9.5
pytz==2023.3
sqlparse==0.4.3
stevedore==5.0.0
virtualenv==20.21.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
